requirements.txt
certifi==2019.11.28
chardet==3.0.4
Cython==0.29.14
docutils==0.15.2
httplib2==0.14.0
idna==2.8
Kivy==1.11.1
Kivy-Garden==0.1.4
pkg-resources==0.0.0
pygame==1.9.6
Pygments==2.5.2
python-congress==0.3.5
requests==2.22.0
six==1.13.0
urllib3==1.25.7
